1) Добавлена библиотека liquibase и авто формирофание таблиц БД 2) Добавлены зависимости Spring security.
