+++ build.gradle
@@ -4,9 +4,9 @@ plugins {
     id 'java'
 }
 
-group = 'ru.javabegin.tasklist'
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '1.8'
+group = 'com.tasklist'
+version = '1.0.1'
+sourceCompatibility = '11'
 
 configurations {
     compileOnly {
@@ -20,8 +20,9 @@ repositories {
 
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springframework.boot:spring-boot-starter-data-rest'
     implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation group: 'org.liquibase', name: 'liquibase-core', version: '4.4.1'
+    implementation 'org.springframework.boot:spring-boot-starter-security'
     compileOnly 'org.projectlombok:lombok'
     developmentOnly 'org.springframework.boot:spring-boot-devtools'
     runtimeOnly 'mysql:mysql-connector-java'
@@ -29,7 +30,6 @@ dependencies {
     testImplementation('org.springframework.boot:spring-boot-starter-test') {
         exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
     }
-    implementation 'org.springframework.boot:spring-boot-starter-security'
 }
 
 test {
