+++ src/main/java/com/termo/tasklist/backendspringboot/BackendSpringbootApplication.java
@@ -1,0 +1,14 @@
+package com.termo.tasklist.backendspringboot;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BackendSpringbootApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BackendSpringbootApplication.class, args);
+    }
+
+}
+
