+++ src/main/java/com/termo/tasklist/backendspringboot/SpringSecurity.java
@@ -1,0 +1,31 @@
+package com.termo.tasklist.backendspringboot;
+
+/**
+ * Created by Termopsis on 29.08.2020.
+ */
+//@SpringBootApplication
+//public class SpringSecurity {
+//
+//    public static void main(String[] args) {
+//        SpringApplication.run(SpringSecurity.class, args);
+//    }
+//
+//}
+
+//@Configuration
+//public class SecurityConfig WebSecurityConfigurerAdapter implements WebMvcConfigurer {
+////    @Override
+////    public void configure(HttpSecurity http) throws Exception {
+////        http.authorizeRequests().antMatchers("/**").permitAll();
+////        http.csrf().disable();
+////
+////    }
+//
+//    @Override
+//    public void addCorsMappings(CorsRegistry registry) {
+//        registry.addMapping("/**")
+//                .allowedOrigins("*")
+//                .allowedMethods("*");
+//    }
+//
+//}
