+++ src/main/java/com/termo/tasklist/backendspringboot/entity/Priority.java
@@ -1,0 +1,44 @@
+package com.termo.tasklist.backendspringboot.entity;
+
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.util.Collection;
+
+/**
+ * Created by Termopsis on 10.08.2020.
+ */
+@Entity
+@NoArgsConstructor
+@Setter
+@EqualsAndHashCode
+public class Priority {
+    private Long id;
+    private String title;
+    private String color;
+    private Collection<Task> tasksById;
+
+    //Аннотация уазывает что поле будет заполняться БД.
+    //Т.к. в бд autoincrement
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    @Column(name = "id")
+    public Long getId() {
+        return id;
+    }
+
+    @Basic
+    @Column(name = "title")
+    public String getTitle() {
+        return title;
+    }
+
+    @Basic
+    @Column(name = "color")
+    public String getColor() {
+        return color;
+    }
+
+}
