+++ src/main/java/com/termo/tasklist/backendspringboot/entity/Stat.java
@@ -1,0 +1,41 @@
+package com.termo.tasklist.backendspringboot.entity;
+
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+/**
+ * Created by Termopsis on 10.08.2020.
+ */
+@Entity
+@NoArgsConstructor
+@Setter
+@EqualsAndHashCode
+public class Stat {
+    private Long id;
+    private Long completedTotal;
+    private Long uncompletedTotal;
+
+    //Только одна запись - новый создаваться не будет
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    public Long getId() {
+        return id;
+    }
+
+    @Basic
+    @Column(name = "completed_total")
+    public Long getCompletedTotal() {
+        return completedTotal;
+    }
+
+    @Basic
+    @Column(name = "uncompleted_total")
+    public Long getUncompletedTotal() {
+        return uncompletedTotal;
+    }
+
+}
