+++ src/main/java/com/termo/tasklist/backendspringboot/search/CategorySearchValues.java
@@ -1,0 +1,19 @@
+package com.termo.tasklist.backendspringboot.search;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+/**
+ * Created by Termopsis on 14.08.2020.
+ */
+@NoArgsConstructor
+@Getter
+@Setter
+@AllArgsConstructor
+public class CategorySearchValues {
+
+    private String title;
+
+}
