--- src/main/java/ru/javabegin/tasklist/backendspringboot/BackendSpringbootApplication.java
@@ -1,14 +1,0 @@
-package ru.javabegin.tasklist.backendspringboot;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BackendSpringbootApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(BackendSpringbootApplication.class, args);
-    }
-
-}
-
