--- src/main/java/ru/javabegin/tasklist/backendspringboot/search/PrioritySearchValues.java
@@ -1,19 +1,0 @@
-package ru.javabegin.tasklist.backendspringboot.search;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-/**
- * Created by Termopsis on 14.08.2020.
- */
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class PrioritySearchValues {
-
-    private String title;
-
-}
