+++ src/main/resources/application.properties
@@ -1,11 +1,16 @@
+server.port=8085
 
-spring.datasource.url = jdbc:mysql://localhost:3306/tasklist?useSSL=false&useUnicode=yes&characterEncoding=UTF-8&allowPublicKeyRetrieval=true&serverTimezone=UTC
-spring.datasource.username = root
-spring.datasource.password = root
-
+spring.datasource.url=jdbc:mysql://localhost:3306/tasklist?createDatabaseIfNotExist=true&useSSL=false&useUnicode=yes&characterEncoding=UTF-8&allowPublicKeyRetrieval=true&serverTimezone=UTC
+spring.datasource.username=root
+spring.datasource.password=root
 
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 
-#Отображать значение параметров переданных в запрос.
+spring.liquibase.change-log=classpath:liquibase/db.changelog-master.xml
+spring.liquibase.url=jdbc:mysql://localhost:3306/tasklist?createDatabaseIfNotExist=true&useSSL=false&useUnicode=yes&characterEncoding=UTF-8&allowPublicKeyRetrieval=true&serverTimezone=UTC
+spring.liquibase.user=root
+spring.liquibase.password=root
+
 logging.level.org.hibernate.type.descriptor.sql=trace
+
