+++ src/test/java/com/termo/tasklist/backendspringboot/BackendSpringbootApplicationTests.java
@@ -1,0 +1,14 @@
+package com.termo.tasklist.backendspringboot;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BackendSpringbootApplicationTests {
+
+    @Test
+    void contextLoads() {
+
+    }
+
+}
