+++ src/test/java/com/termo/tasklist/backendspringboot/service/CategoryServiceTest.java
@@ -1,0 +1,17 @@
+package com.termo.tasklist.backendspringboot.service;
+
+import org.junit.jupiter.api.Test;
+
+/**
+ * Created by Termopsis on 06.09.2020.
+ */
+class CategoryServiceTest {
+
+    @Test
+    void deleteById() {
+
+
+        //Assert
+
+    }
+}
