--- src/test/java/ru/javabegin/tasklist/backendspringboot/BackendSpringbootApplicationTests.java
@@ -1,14 +1,0 @@
-package ru.javabegin.tasklist.backendspringboot;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BackendSpringbootApplicationTests {
-
-    @Test
-    void contextLoads() {
-
-    }
-
-}
