--- src/test/java/ru/javabegin/tasklist/backendspringboot/service/CategoryServiceTest.java
@@ -1,17 +1,0 @@
-package ru.javabegin.tasklist.backendspringboot.service;
-
-import org.junit.jupiter.api.Test;
-
-/**
- * Created by Termopsis on 06.09.2020.
- */
-class CategoryServiceTest {
-
-    @Test
-    void deleteById() {
-
-
-        //Assert
-
-    }
-}
